+++ src/lantern/exporters/site.py
@@ -13,7 +13,7 @@ from lantern.exporters.base import Exporter as BaseExporter
 from lantern.exporters.records import RecordsExporter
 from lantern.exporters.waf import WebAccessibleFolderExporter
 from lantern.exporters.website import WebsiteSearchExporter
-from lantern.models.site import ExportMeta
+from lantern.models.site import ExportMeta, SitePageMeta
 from lantern.stores.base import SelectRecordsProtocol, StoreInitProtocol
 from lantern.utils import get_jinja_env, get_record_aliases, prettify_html
 
@@ -242,21 +242,33 @@ class SitePagesExporter(Exporter):
         """Initialise exporter."""
         super().__init__(logger=logger, meta=meta, s3=s3)
         self._jinja = get_jinja_env()
-        self._templates = [
-            "_views/404.html.j2",
-            "_views/legal/accessibility.html.j2",
-            "_views/legal/cookies.html.j2",
-            "_views/legal/copyright.html.j2",
-            "_views/legal/privacy.html.j2",
-            "_views/-/formatting.html.j2",
-        ]
-        self._html_title = {
-            "_views/404.html.j2": "Not Found",
-            "_views/legal/accessibility.html.j2": "Accessibility Statement",
-            "_views/legal/cookies.html.j2": "Cookies Policy",
-            "_views/legal/copyright.html.j2": "Copyright Policy",
-            "_views/legal/privacy.html.j2": "Privacy Policy",
-            "_views/-/formatting.html.j2": "Supported Formatting Guide",
+        self._templates: dict[str, SitePageMeta] = {
+            "_views/404.html.j2": SitePageMeta(title="Not Found", url="-", meta=False),
+            "_views/legal/accessibility.html.j2": SitePageMeta(
+                title="Accessibility Statement",
+                url=f"{self._meta.base_url}/legal/accessibility",
+                description="Basic accessibility check for BAS Data Catalogue",
+            ),
+            "_views/legal/cookies.html.j2": SitePageMeta(
+                title="Cookies Policy",
+                url=f"{self._meta.base_url}/legal/cookies",
+                description="Cookies policy for BAS Data Catalogue",
+            ),
+            "_views/legal/copyright.html.j2": SitePageMeta(
+                title="Copyright Policy",
+                url=f"{self._meta.base_url}/legal/copyright",
+                description="Copyright policy for BAS Data Catalogue",
+            ),
+            "_views/legal/privacy.html.j2": SitePageMeta(
+                title="Privacy Policy",
+                url=f"{self._meta.base_url}/legal/privacy",
+                description="Privacy policy for BAS Data Catalogue",
+            ),
+            "_views/-/formatting.html.j2": SitePageMeta(
+                title="Formatting Guide",
+                url=f"{self._meta.base_url}/-/formatting",
+                description="Formatting guide for freetext record elements",
+            ),
         }
 
     def _get_page_path(self, template_path: str) -> Path:
@@ -267,7 +279,10 @@ class SitePagesExporter(Exporter):
 
     def _dumps(self, template_path: str) -> str:
         """Build a page."""
-        self._meta.html_title = self._html_title[template_path]
+        page_meta: SitePageMeta = self._templates[template_path]
+        self._meta.html_title = page_meta.title
+        self._meta.html_open_graph = page_meta.open_graph
+        self._meta.html_schema_org = page_meta.schema_org
         raw = self._jinja.get_template(template_path).render(meta=self._meta)
         return prettify_html(raw)