+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 * Disabling lineage statement in container item types
 * Missing datestamp update in the clone record development task
 * Correcting names for `STORE_GITLAB_CACHE_PATH`, `STORE_GITLAB_ENDPOINT` and `STORE_GITLAB_PROJECT_ID` config options
+* Missing Open Graph and Schema.org metadata for static site pages
 
 ### Changed
 
@@ -56,6 +57,7 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 * Reorganising exporter utility methods
 * Refactoring inline JavaScript in templates into standalone files
 * Refactored CSS build process consistency with new 'js' process, inc. renaming 'tailwind' development task to 'css'
+* Extracted Open Graph and Schema.org link preview metadata logic out of catalogue items into site data classes
 
 ### Removed